refactor(storage): Move write_batch::codec to a separate file
